+++ src/hu/banattila/modellek/emberek/Konyvelo.java
@@ -10,11 +10,11 @@ public class Konyvelo extends Szemelyzet {
     }
 
     @Override
-    public void setFizetes(double napiToke) {
-        this.fizetes = napiToke * 0.02;
+    public void setFizetes(int napiToke) {
+        this.fizetes = (int)(napiToke * 0.02);
     }
 
-    public double getFizetes(){
+    public int getFizetes(){
         return this.fizetes;
     }
